tmux 3.2 has started to set TERM_PROGRAM to "tmux" (#84)
This causes zsh-notify to silently stop working on macOS, because it
assumes TERM_PROGRAM to be set to `iTerm2.app` or `Apple_Terminal`.

So, we should make zsh-notify smarter in how it detects iTerm2.
Unfortunately, it doesn't seem completely trivial to do that.
The official way seems to be this rather long script
https://iterm2.com/utilities/it2check

In this PR I suggest a simplistic method of  checking ITERM_SESSION_ID.
That seems to be set even inside tmux, as long as the tmux
session was started from iTerm2.

TERM_SESSION_ID is similarly for Apple Terminal.

I don't have an arsenal of machines to test
this on. So, if people report where it fails, we can improve it.