Use $EPOCHSECONDS instead of $(date)
This switches to getting the current time from the shell variable
`$EPOCHSECONDS` rather than executing the `date` command. This results
in the average time of the combined before and after functions going
from 10ms to 6ms on my machine.

The `$EPOCHSECONDS` variable was added to the Zsh codebase on 2003-10-08
and appears to have been released in Zsh 4.2.0 on 2004-03-19 (verified
by looking at the Zsh 4.2.0 release tarball).